// extensions/advent-shipping/generated/api.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * Represents an [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601)-encoded date string.
   * For example, September 7, 2019 is represented as `"2019-07-16"`.
   */
  Date: any;
  /**
   * Represents an [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601)-encoded date and time string.
   * For example, 3:50 pm on September 7, 2019 in the time zone of UTC (Coordinated Universal Time) is
   * represented as `"2019-09-07T15:50:00Z`".
   */
  DateTime: any;
  /**
   * A subset of the [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format that
   * includes the date and time but not the timezone which is determined from context.
   *
   * For example, "2018-01-01T00:00:00".
   */
  DateTimeWithoutTimezone: any;
  /**
   * A signed decimal number, which supports arbitrary precision and is serialized as a string.
   *
   * Example values: `"29.99"`, `"29.999"`.
   */
  Decimal: any;
  /**
   * A function-scoped handle to a refer a resource.
   * The Handle type appears in a JSON response as a String, but it is not intended to be human-readable.
   * Example value: `"10079785100"`
   */
  Handle: any;
  /**
   * A [JSON](https://www.json.org/json-en.html) object.
   *
   * Example value:
   * `{
   *   "product": {
   *     "id": "gid://shopify/Product/1346443542550",
   *     "title": "White T-shirt",
   *     "options": [{
   *       "name": "Size",
   *       "values": ["M", "L"]
   *     }]
   *   }
   * }`
   */
  JSON: any;
  /**
   * A subset of the [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format that
   * includes the time but not the date or timezone which is determined from context.
   * For example, "05:43:21".
   */
  TimeWithoutTimezone: any;
  /** A void type that can be used to return a null value from a mutation. */
  Void: any;
};

/** Represents a generic custom attribute. */
export type Attribute = {
  __typename?: 'Attribute';
  /** Key or name of the attribute. */
  key: Scalars['String'];
  /** Value of the attribute. */
  value?: Maybe<Scalars['String']>;
};

/** Represents information about the buyer that is interacting with the cart. */
export type BuyerIdentity = {
  __typename?: 'BuyerIdentity';
  /** The customer associated with the cart. */
  customer?: Maybe<Customer>;
  /** The email address of the buyer that's interacting with the cart. */
  email?: Maybe<Scalars['String']>;
  /** Whether the buyer authenticated with a customer account. */
  isAuthenticated: Scalars['Boolean'];
  /** The phone number of the buyer that's interacting with the cart. */
  phone?: Maybe<Scalars['String']>;
  /** The purchasing company associated with the cart. */
  purchasingCompany?: Maybe<PurchasingCompany>;
};

/** A cart represents the merchandise that a buyer intends to purchase, and the cost associated with the cart. */
export type Cart = {
  __typename?: 'Cart';
  /** The attributes associated with the cart. Attributes are represented as key-value pairs. */
  attribute?: Maybe<Attribute>;
  /** Information about the buyer that is interacting with the cart. */
  buyerIdentity?: Maybe<BuyerIdentity>;
  /** The costs that the buyer will pay at checkout. */
  cost: CartCost;
  /** A list of lines containing information about the items that can be delivered. */
  deliverableLines: Array<DeliverableCartLine>;
  /** The delivery groups available for the cart based on the buyer's shipping address. */
  deliveryGroups: Array<CartDeliveryGroup>;
  /** A list of lines containing information about the items the customer intends to purchase. */
  lines: Array<CartLine>;
};


/** A cart represents the merchandise that a buyer intends to purchase, and the cost associated with the cart. */
export type CartAttributeArgs = {
  key?: InputMaybe<Scalars['String']>;
};

/** The cost that the buyer will pay at checkout. */
export type CartCost = {
  __typename?: 'CartCost';
  /** The amount, before taxes and discounts, for the customer to pay. */
  subtotalAmount: MoneyV2;
  /** The total amount for the customer to pay. */
  totalAmount: MoneyV2;
  /** The duty amount for the customer to pay at checkout. */
  totalDutyAmount?: Maybe<MoneyV2>;
  /** The tax amount for the customer to pay at checkout. */
  totalTaxAmount?: Maybe<MoneyV2>;
};

/** Information about the options available for one or more line items to be delivered to a specific address. */
export type CartDeliveryGroup = {
  __typename?: 'CartDeliveryGroup';
  /** A list of cart lines for the delivery group. */
  cartLines: Array<CartLine>;
  /** The destination address for the delivery group. */
  deliveryAddress?: Maybe<MailingAddress>;
  /** The delivery options available for the delivery group. */
  deliveryOptions: Array<CartDeliveryOption>;
  /** Unique identifier for the delivery group. */
  id: Scalars['ID'];
  /** Information about the delivery option the buyer has selected. */
  selectedDeliveryOption?: Maybe<CartDeliveryOption>;
};

/** Information about a delivery option. */
export type CartDeliveryOption = {
  __typename?: 'CartDeliveryOption';
  /** The code of the delivery option. */
  code?: Maybe<Scalars['String']>;
  /** The cost for the delivery option. */
  cost: MoneyV2;
  /** The method for the delivery option. */
  deliveryMethodType: DeliveryMethod;
  /** The description of the delivery option. */
  description?: Maybe<Scalars['String']>;
  /** The unique identifier of the delivery option. */
  handle: Scalars['Handle'];
  /** The title of the delivery option. */
  title?: Maybe<Scalars['String']>;
};

/** Represents information about the merchandise in the cart. */
export type CartLine = {
  __typename?: 'CartLine';
  /**
   * Retrieve a cart line attribute by key.
   *
   * Cart line attributes are also known as line item properties in Liquid.
   */
  attribute?: Maybe<Attribute>;
  /** The cost of the merchandise line that the buyer will pay at checkout. */
  cost: CartLineCost;
  /** The ID of the cart line. */
  id: Scalars['ID'];
  /** The merchandise that the buyer intends to purchase. */
  merchandise: Merchandise;
  /** The quantity of the merchandise that the customer intends to purchase. */
  quantity: Scalars['Int'];
  /**
   * The selling plan associated with the cart line and the effect that each
   * selling plan has on variants when they're purchased.
   */
  sellingPlanAllocation?: Maybe<SellingPlanAllocation>;
};
// ...
export type CustomProduct = {
  __typename?: 'CustomProduct';
  /** Whether the merchandise is a gift card. */
  isGiftCard: Scalars['Boolean'];
  /** Whether the merchandise requires shipping. */
  requiresShipping: Scalars['Boolean'];
  /** The localized title of the product in the customer’s locale. */
  title: Scalars['String'];
  /** The weight of the product variant in the unit system specified with `weight_unit`. */
  weight?: Maybe<Scalars['Float']>;
  /** Unit of measurement for weight. */
  weightUnit: WeightUnit;
};
// ...
export type DeliverableCartLine = {
  __typename?: 'DeliverableCartLine';
  /**
   * Retrieve a cart line attribute by key.
   *
   * Cart line attributes are also known as line item properties in Liquid.
   */
  attribute?: Maybe<Attribute>;
  /** The ID of the cart line. */
  id: Scalars['ID'];
  /** The merchandise that the buyer intends to purchase. */
  merchandise: Merchandise;
  /** The quantity of the merchandise that the customer intends to purchase. */
  quantity: Scalars['Int'];
};
// ...
export type Discount = {
  /** The discount message. */
  message?: InputMaybe<Scalars['String']>;
  /**
   * The targets of the discount.
   *
   * This argument accepts a collection of either `ProductVariantTarget`s or `CartLineTarget`s, but not both.
   */
  targets: Array<Target>;
  /** The value of the discount. */
  value: Value;
};

/** The strategy that's applied to the list of discounts. */
export enum DiscountApplicationStrategy {
  /** Apply all discounts with conditions that are satisfied. This does not override discount combination or stacking rules. */
  All = 'ALL',
  /** Only apply the first discount with conditions that are satisfied. */
  First = 'FIRST',
  /** Only apply the discount that offers the maximum reduction. */
  Maximum = 'MAXIMUM'
}

/** A discount wrapper node. */
export type DiscountNode = HasMetafields & {
  __typename?: 'DiscountNode';
  /** Returns a metafield by namespace and key that belongs to the resource. */
  metafield?: Maybe<Metafield>;
};
// ...
export type FunctionRunResult = {
  /** The strategy to apply the list of discounts. */
  discountApplicationStrategy: DiscountApplicationStrategy;
  /** The list of discounts to be applied. */
  discounts: Array<Discount>;
};
// ...
export type HasTagResponse = {
  __typename?: 'HasTagResponse';
  /** Whether the current object has the tag. */
  hasTag: Scalars['Boolean'];
  /** The tag. */
  tag: Scalars['String'];
};

/** The input object for the function. */
export type Input = {
  __typename?: 'Input';
  /** The cart. */
  cart: Cart;
  /** The discount node executing the function. */
  discountNode: DiscountNode;
  /** The localization of the Function execution context. */
  localization: Localization;
  /** The conversion rate between the shop's currency and the currency of the cart. */
  presentmentCurrencyRate: Scalars['Decimal'];
  /** Information about the shop. */
  shop: Shop;
};
// ...
export type Localization = {
  __typename?: 'Localization';
  /** The country of the active localized experience. */
  country: Country;
  /** The language of the active localized experience. */
  language: Language;
  /** The market of the active localized experience. */
  market: Market;
};
// ...
export type Merchandise = CustomProduct | ProductVariant;
// ...
export type MoneyV2 = {
  __typename?: 'MoneyV2';
  /** Decimal money amount. */
  amount: Scalars['Decimal'];
  /** Currency of the money. */
  currencyCode: CurrencyCode;
};
// ...
export type Product = HasMetafields & {
  __typename?: 'Product';
  /** A unique human-friendly string of the product's title. */
  handle: Scalars['Handle'];
  /** Whether the product has any of the given tags. */
  hasAnyTag: Scalars['Boolean'];
  /** Whether the product has the given tags. */
  hasTags: Array<HasTagResponse>;
  /** A globally-unique identifier. */
  id: Scalars['ID'];
  /** Whether the product is in any of the given collections. */
  inAnyCollection: Scalars['Boolean'];
  /** Whether the product is in the given collections. */
  inCollections: Array<CollectionMembership>;
  /** Whether the product is a gift card. */
  isGiftCard: Scalars['Boolean'];
  /** Returns a metafield by namespace and key that belongs to the resource. */
  metafield?: Maybe<Metafield>;
  /** The product type specified by the merchant. */
  productType?: Maybe<Scalars['String']>;
  /** The localized title of the product in the customer’s locale. */
  title: Scalars['String'];
  /** The name of the product's vendor. */
  vendor?: Maybe<Scalars['String']>;
};
// ...
export type ProductVariant = HasMetafields & {
  __typename?: 'ProductVariant';
  /** A globally-unique identifier. */
  id: Scalars['ID'];
  /** Returns a metafield by namespace and key that belongs to the resource. */
  metafield?: Maybe<Metafield>;
  /** The product that this variant belongs to. */
  product: Product;
  /** Whether the merchandise requires shipping. */
  requiresShipping: Scalars['Boolean'];
  /** An identifier for the product variant in the shop. Required in order to connect to a fulfillment service. */
  sku?: Maybe<Scalars['String']>;
  /** The localized title of the product variant in the customer’s locale. */
  title?: Maybe<Scalars['String']>;
  /** The weight of the product variant in the unit system specified with `weight_unit`. */
  weight?: Maybe<Scalars['Float']>;
  /** Unit of measurement for weight. */
  weightUnit: WeightUnit;
};
// ...
export type Shop = HasMetafields & {
  __typename?: 'Shop';
  /** Information about the current time relative to the shop's timezone setting. */
  localTime: LocalTime;
  /** Returns a metafield by namespace and key that belongs to the resource. */
  metafield?: Maybe<Metafield>;
};
// ...
export type RunInput = { __typename?: 'Input', cart: { __typename?: 'Cart', lines: Array<{ __typename?: 'CartLine', id: string, quantity: number, merchandise: { __typename: 'CustomProduct' } | { __typename: 'ProductVariant', id: string, product: { __typename?: 'Product', title: string, hasTags: Array<{ __typename?: 'HasTagResponse', tag: string, hasTag: boolean }> } } }>, deliveryGroups: Array<{ __typename?: 'CartDeliveryGroup', id: string, deliveryOptions: Array<{ __typename?: 'CartDeliveryOption', handle: any, title?: string | null, cost: { __typename?: 'MoneyV2', amount: any } }> }> } };